Dynamic Dates for addcomments
The default was to have every comment dated on Feb 26th, but for filter page to work new comments need to be represented of when they were posted. Created $time to reflect dynamic date on each new post to use in filter.php, also uploading image for calender for filter.php when filter is completed.